util/f32_to_uf11: make use of optional value range
In the CTS test 'KHR-GL45.copy_image.functional' we ended up doing a sw format
conversion from R11F_G11F_B10F_EXT to RGBA back to R11F_G11F_B10F_EXT within
glGetTexImage calls.

By total misfortune we also got f11 values of 0x4, which got converted to a
0x36800000 f32 value in uf11_to_f32. But the result of f32_to_uf11(0x36800000)
was 0x0 which led to the above mentioned test to fail due to invalid data read
by glGetTexImage.

Adjusting f32_to_uf11 to make use of the optional value range for
	E == 0 and M != 0
so that we have more 'x' for which is the following is true
	f32_to_uf11(uf11_to_f32(x)) == x

Fixes a few subtests for KHR-GL45.copy_image.functional on Nouveau and i965.

Signed-off-by: Karol Herbst <karolherbst@gmail.com>