WIP nvc0/ir: change textureGrad to always use lane 0 as the tex origin
TODO: why is this needed?
TODO: gm107

Thanks to Karol Herbst for the debugging / tracing work that led to this
change.

Move to using lane 0 as the "work" lane for the texture. It is unclear
why this helps, as that computation should be identical to doing it in
the "correct" lane with the properly adjusted quadops.

In order to be able to use the lane 0 result, we also have to ensure
that lane 0 contains the proper array/indirect values.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>