WIP: clover: add support for nir shaders via spirv
For drivers that want PIPE_SHADER_IR_NIR, translate spirv to nir at
kernel creation time.

This builds on spirv support, and continues to use spirv as the
serialization format, since that should be a more stable ABI.  (Yes, it
is possible to serialize nir, but that is not expected to be a stable
binary format across mesa versions.)

TODO some of this could probably be cleaner?

TODO need to ralloc_free() the nir somewhere..