WIP: nir: add deref_ptr_as_array instruction
This matches the behaviour of the 'Element' parameter in the spirv
Op*PtrAccessChain instructions, handling the pointer dereference as
an array dereference without shedding an array type (ie. the result
has the same glsl_type)

To think of it in c/cl terms, "ptr->" or "(*ptr)." are the same
thing as "ptr[0].".