WIP: nir: add back nir_var_param
Possibly squash into "nir,spirv: Rework function calls".  Or possibly go
back to having glsl_type in the nir_function, and nir_variable in the
nir_function_impl.

(Tbh, var->data.driver_loc ends up defining the ABI for the function, so
this seems like it belongs in the function prototype, not the function
body.)