clover+gallium: support per-kernel limits
Some limits, such as max # of threads in a work-group, vary depending
on the resources (ie. registers) used by a kernel.  OpenCL provides
clGetKernelWorkGroupInfo() for querying these kernel specific limits.
To implement this properly, we need a variant of get_compute_param()
which takes the compute-state CSO as an argument.

Signed-off-by: Rob Clark <robdclark@gmail.com>