clover+gallium+freedreno: caps to reduce kernel recompiles
Not all drivers care when cs.reg_*_mem change.  (ir3 only cares about
req_input_mem and removing that dependency should be easy.)  Add some
caps to let clover make better decisions about when it needs to re-
create the compute-state CSO.

This way, if the kernel is compiled early for clGetKernelWorkGroupInfo()
it doesn't end up getting compiled a second time when the kernel is
launched for the first time (clEnqueueNDRangeKernel(), etc).

Signed-off-by: Rob Clark <robdclark@gmail.com>