nir: add all combinations of conversions with rounding and saturation
OpenCL has explicit casts where one can specify the rounding mode and put a
sat modifier:

https://www.khronos.org/registry/OpenCL/sdk/2.1/docs/man/xhtml/convert_T.html

_sat is valid for all conversions to an integer type and rounding modes are
valid for all conversions involving floats.

Allthough the FPRoundingMode modifier is allowed without any restrictions in
capabilities, it can only be used together with fp16 in GLSL. Additionally it
can be used for conversions to/from floating points in OpenCL.

The SaturatedConversion modifier, OpSatConvertUToS and OpSatConvertSToU are
only supported for Kernels, so current drivers are safe.

Signed-off-by: Karol Herbst <kherbst@redhat.com>