nir: Add load_ssbo_address/size intrinsics
To facilitate lowering SSBOs to globals, we need a load_ssbo_address
intrinsic. This intrinsic takes an SSBO index and loads the address in
global memory of the SSBO (likely implemented via a uniform in the
driver). Further, to support bounds checking during lowering, we have a
load_ssbo_size intrinsic used as the upper bound on offsets (again,
likely implemented with a uniform).

Signed-off-by: Alyssa Rosenzweig <alyssa.rosenzweig@collabora.com>