nv50/ir/nir: insert joins for loops as well
This is required for Volta as we have a shader managed stack and need to be
careful about the barrier operations. Making sure all threads converged
before executing joinats and joins is the most reliable way of doing that.

Without this change we could have a loop between ifs diverging threads
until they hit the next joinat.

Signed-off-by: Karol Herbst <kherbst@redhat.com>